refactoring parallel execution queues
